requirements.txt
streamlit
google-generativeai
supabase
asyncio
deepgram-sdk